Moved output to a separate module, shifting a utiltity out of main
