+++ src/main.ts
@@ -1,46 +1,14 @@
 import { Type } from '@freik/core-utils';
-import { promises as fs } from 'fs';
 import minimist from 'minimist';
 import path from 'path';
 import { EnumerateBoards } from './board.js';
 import { IsConfigPresent, ReadConfig } from './config.js';
+import { Dump, FlushOutput, SetOutputFile } from './dump.js';
 import { MakeGlobals } from './globals.js';
 import { GetLibraries } from './libraries.js';
 import { ParseFile } from './parser.js';
 import { GetGnuMakeTarget } from './targets/gnumake.js';
-import { Definition, PlatformTarget, Recipe } from './types.js';
-
-let outputFile: string[] | undefined;
-let outputName: string | undefined;
-
-function setOutputFile(dest: unknown) {
-  if (Type.isString(dest)) {
-    outputName = dest;
-    outputFile = [];
-  }
-}
-
-function dumpToFile(message: unknown): void {
-  if (outputFile !== undefined && Type.isString(message)) {
-    outputFile.push(message);
-  }
-}
-
-// Eventually, we can dump stuff into different files, right?
-export function Dump(which?: string): (message: unknown) => void {
-  switch (which) {
-    case undefined:
-    case 'log':
-      return Type.isUndefined(outputFile) ? console.log : dumpToFile; // eslint-disable-line no-console
-    case 'err':
-      return console.error; // eslint-disable-line no-console
-    default:
-      return (msg) => {
-        // eslint-disable-next-line no-console
-        console.error(which, ' is an invalid selector: ', msg);
-      };
-  }
-}
+import { PlatformTarget } from './types.js';
 
 // Overall structure:
 // Walk the platform.txt file, documented here:
@@ -92,7 +60,7 @@ async function parseCommandLine(args: string[]): Promise<string[]> {
   if (Type.hasStr(argv, 'config')) {
     await ReadConfig(argv.config);
   }
-  setOutputFile(argv?.out);
+  SetOutputFile(argv?.out);
   if (Type.hasStr(argv, 'target')) {
     if (argv.target.toLocaleLowerCase() !== 'gnumake') {
       throw Error(`Command line error: Unsupported target ${argv.target}`);
@@ -127,9 +95,7 @@ export default async function main(...args: string[]): Promise<void> {
     await platformTarget.emit(platformPath, platSyms, boardSyms, libraries);
 
     // Flush the output to disk...
-    if (!Type.isUndefined(outputFile) && Type.isString(outputName)) {
-      await fs.writeFile(outputName, outputFile.join('\n'), 'utf-8');
-    }
+    await FlushOutput();
   } catch (e) {
     const name = Type.hasStr(e, 'name') ? e.name : '<unknown>';
     const message = Type.hasStr(e, 'message') ? e.message : '<no message>';
@@ -140,65 +106,4 @@ export default async function main(...args: string[]): Promise<void> {
         : -1;
     ShowHelp(`Error: ${name} @ ${file}#${line}:\n${message}`);
   }
-}
-
-export function CalculateChecksAndOrderDefinitions(
-  defs: Definition[],
-  rules: Recipe[],
-  optionalDefs: string[],
-): { checks: string[]; defs: Definition[] } {
-  // Don't know if I'll need the rules or not
-
-  // First, let's identify all the mandatory user-defined symbols
-  const allDefs = new Set(defs.map((d) => d.name));
-  const tmp: string[] = [];
-  const allDeps = new Set(
-    tmp.concat(
-      ...defs.map((d) => d.dependsOn),
-      ...rules.map((rec) => rec.dependsOn),
-    ),
-  );
-  // Remove allDefs from allDeps
-  const checks: Set<string> = new Set(
-    [...allDeps].filter((x) => !allDefs.has(x)),
-  );
-  const done: Set<string> = new Set(checks);
-  // Clear out known optional values
-  optionalDefs.forEach((a) => checks.delete(a));
-
-  // Now checks has the list of all undefined symbols
-  // These should have checks emitted in the makefile to validate that the user
-  // has defined them already (or have defaults assigned if that's unimportant)
-
-  function allDefined(def: Definition): boolean {
-    for (const d of def.dependsOn) {
-      if (!done.has(d)) {
-        return false;
-      }
-    }
-    return true;
-  }
-  const ordered: Definition[] = [];
-  const stillPending: Map<string, number> = new Map();
-  defs.forEach((d: Definition) => {
-    let val = stillPending.get(d.name);
-    if (typeof val !== 'number') {
-      val = 0;
-    }
-    stillPending.set(d.name, val + 1);
-  });
-  // This is such a lame, slow sorting algorithm. I really should do better...
-  const skip: Set<number> = new Set();
-  for (let i = 0; i < defs.length; i++) {
-    if (!skip.has(i)) {
-      if (allDefined(defs[i])) {
-        ordered.push(defs[i]);
-        done.add(defs[i].name);
-        skip.add(i);
-        i = -1;
-      }
-    }
-  }
-
-  return { checks: [...checks.keys()], defs: ordered };
 }
