+++ src/utils.ts
@@ -1,7 +1,13 @@
 import { Type } from '@freik/core-utils';
 import { GetTarget } from './main.js';
 import { LookupSymbol } from './symbols.js';
-import type { DependentValue, ParsedFile, SimpleSymbol } from './types.js';
+import type {
+  Definition,
+  DependentValue,
+  ParsedFile,
+  Recipe,
+  SimpleSymbol,
+} from './types.js';
 
 // This takes a value, and returns the resolved value plus the list of
 // undefined names within the value
@@ -108,4 +114,65 @@ export function Unquote(inv: string): string {
     return inv;
   }
   return inv.substring(1, inv.length - 1);
+}
+
+export function CalculateChecksAndOrderDefinitions(
+  defs: Definition[],
+  rules: Recipe[],
+  optionalDefs: string[],
+): { checks: string[]; defs: Definition[] } {
+  // Don't know if I'll need the rules or not
+
+  // First, let's identify all the mandatory user-defined symbols
+  const allDefs = new Set(defs.map((d) => d.name));
+  const tmp: string[] = [];
+  const allDeps = new Set(
+    tmp.concat(
+      ...defs.map((d) => d.dependsOn),
+      ...rules.map((rec) => rec.dependsOn),
+    ),
+  );
+  // Remove allDefs from allDeps
+  const checks: Set<string> = new Set(
+    [...allDeps].filter((x) => !allDefs.has(x)),
+  );
+  const done: Set<string> = new Set(checks);
+  // Clear out known optional values
+  optionalDefs.forEach((a) => checks.delete(a));
+
+  // Now checks has the list of all undefined symbols
+  // These should have checks emitted in the makefile to validate that the user
+  // has defined them already (or have defaults assigned if that's unimportant)
+
+  function allDefined(def: Definition): boolean {
+    for (const d of def.dependsOn) {
+      if (!done.has(d)) {
+        return false;
+      }
+    }
+    return true;
+  }
+  const ordered: Definition[] = [];
+  const stillPending: Map<string, number> = new Map();
+  defs.forEach((d: Definition) => {
+    let val = stillPending.get(d.name);
+    if (typeof val !== 'number') {
+      val = 0;
+    }
+    stillPending.set(d.name, val + 1);
+  });
+  // This is such a lame, slow sorting algorithm. I really should do better...
+  const skip: Set<number> = new Set();
+  for (let i = 0; i < defs.length; i++) {
+    if (!skip.has(i)) {
+      if (allDefined(defs[i])) {
+        ordered.push(defs[i]);
+        done.add(defs[i].name);
+        skip.add(i);
+        i = -1;
+      }
+    }
+  }
+
+  return { checks: [...checks.keys()], defs: ordered };
 }
